# sweep line perf