--- a/geo/benches/utils/crossings.rs
+++ b/geo/benches/utils/crossings.rs
@@ -1,14 +1,12 @@
 #![allow(dead_code)]
 
-use std::iter::FromIterator;
-
 use geo::algorithm::sweep::Intersections;
 use geo::{line_intersection::line_intersection, Line};
 
 use rstar::{primitives::GeomWithData, RTree};
 
 pub fn count_bo(lines: Vec<Line<f64>>) -> usize {
-    Intersections::from_iter(lines).count()
+    Intersections::from_iter(lines).iter().count()
 }
 
 pub fn count_brute(lines: &[Line<f64>]) -> usize {