--- a/geo/src/algorithm/old_sweep/iter.rs
+++ b/geo/src/algorithm/old_sweep/iter.rs
@@ -180,7 +180,7 @@ where
 /// See the trait documentation for more information on usage with
 /// custom types.
 ///
-/// ```rust
+/// ```rust,ignore
 /// use geo::Line;
 /// use geo::sweep::Intersections;
 /// use std::iter::FromIterator;