@urschrei 2025-07-17
> in particular I feel like talking about "events" sort of muddies the waters.

Well, they're still present and referred to by name in the implementation…

The docs as written were as much a way for me to understand how the algorithm works as anything else, and the B-O algorithm specifically uses "event" as way of discussing a specific step in the algorithm. I'm in no way wedded to that term, but it's – IMO – important to be able to explain how this implementation's Rust-specific moving parts relate to the "classic" algorithm, because it provides a way for people who may not have much Rust knowledge, but are familiar with the algorithm to find a way in (this difficulty is essentially what happens every time I look at JTS, and even more so for GEOS).

So as written, your current draft description makes sense, but I think it'd be even better if it could point at B-O where possible. How do we get from your explanation to "sweep line" and "event"?