@michaelkirk 2025-07-17
Ok, I've have a go at reconciling the documentation with BO. Please take another look @urschrei